+++ Engine - copia/EngineLib/TileMap.h
@@ -13,7 +13,7 @@ public:
 	DllExport void setTile(const Tile& tile);
 	DllExport void setMapTileId(int layer, unsigned int column, unsigned int row, unsigned int id);
 
-	DllExport void setDimensions(float width, float heigth);
+	DllExport void setDimensions(unsigned int width, unsigned int heigth);
 	DllExport void setTileDimensions(float tileWidth, float tileHeigth);
 
 	DllExport void setTexture(const Texture& texture);
